microservicio_ia/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
# --- VERSIONES ESTABLES 0.2 (Crucial para que no falle) ---
langchain-core==0.2.38
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
# ----------------------------------------------------------
google-generativeai
chromadb
pypdf
python-multipart
